fix : reorganize notification types and payload structure and fix sendnotification method
