+++ api/src/notifications/dto/notification-payload.ts
@@ -1,10 +1,5 @@
 import { IsEnum, IsNotEmpty, IsOptional, IsString } from 'class-validator';
-
-export enum NotificationType {
-  FRIEND_REQUEST = 'FRIEND_REQUEST',
-  NEW_ACHIEVEMENT = 'NEW_ACHIEVEMENT',
-  GAME_INVITE = 'GAME_INVITE',
-}
+import { NotificationType } from '../Constants/notification-type.enum';
 
 export class NotificationPayload {
   @IsEnum(NotificationType)
